pyproject.toml
[project]
name = "apps"
version = "0.1.0"
description = "General Apps and Calculators"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi",
    "marimo",
    "starlette",
    "jinja2",
    "itsdangerous",
    "python-dotenv",
    "python-multipart",
    "passlib",
    "pydantic",
    "vega-datasets==0.9.0",
    "humanize",
    "pandas",
    "requests",
    "numpy_financial",
    "matplotlib",
]
